+++ storybook/lib/stories/list/list_tile.dart
@@ -8,41 +8,42 @@ final Story listTileStory = Story(
   name: 'General/List/List tile',
   builder: (context) {
     final k = context.knobs;
+    final title = k.text(label: 'Title', initial: 'Title');
+    final subtitle = k.text(label: 'Subtitle', initial: 'Subtitle');
+    final prefix = k.options(
+      label: 'Prefix',
+      initial: null,
+      options: exampleIcons,
+    );
+    final suffix = k.options(
+      label: 'Suffix',
+      initial: null,
+      options: exampleIcons,
+    );
+    final info = k.text(label: 'Info', initial: 'Info');
+    final infoWidget = k.options(
+      label: 'Info widget',
+      initial: null,
+      options: exampleIcons,
+    );
+    final fontVariant = k.options(
+      label: 'Font variant',
+      initial: FontVariant.normal,
+      options: FontVariant.values.toOptions(),
+    );
 
     return SingleChildScrollView(
       child: Column(
         children: Iterable<int>.generate(10)
             .map(
               (i) => OptimusListTile(
-                title: Text(k.text(label: 'Title', initial: 'Title')),
-                subtitle: Text(k.text(label: 'Subtitle', initial: 'Subtitle')),
-                info: Text(k.text(label: 'Info', initial: 'Info')),
-                fontVariant: k.options(
-                  label: 'Font variant',
-                  initial: FontVariant.normal,
-                  options: FontVariant.values.toOptions(),
-                ),
-                suffix: Icon(
-                  k.options(
-                    label: 'Suffix',
-                    initial: null,
-                    options: exampleIcons,
-                  ),
-                ),
-                prefix: Icon(
-                  k.options(
-                    label: 'Prefix',
-                    initial: null,
-                    options: exampleIcons,
-                  ),
-                ),
-                infoWidget: Icon(
-                  k.options(
-                    label: 'Info widget',
-                    initial: null,
-                    options: exampleIcons,
-                  ),
-                ),
+                title: Text(title),
+                subtitle: Text(subtitle),
+                info: Text(info),
+                fontVariant: fontVariant,
+                prefix: prefix != null ? Icon(prefix) : null,
+                suffix: suffix != null ? Icon(suffix) : null,
+                infoWidget: infoWidget != null ? Icon(infoWidget) : null,
                 onTap: () {},
               ),
             )