+++ remote_logger/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.0+1
+
+ - **REFACTOR**: Bump mews_pedantic to 0.13.0 (#319).
+ - **REFACTOR**: Fix new lint errors.
+
+## 0.2.0
+
+> Note: This release has breaking changes.
+
+ - **BREAKING** **REFACTOR**: Updated to Flutter 3.7.
+
 ## 0.1.0+2
 
  - **REFACTOR**: Follow new mews_pedantic rules.